🚚 move: DashboardSkeleton device 폴더 분리
