+++ app/src/components/templates/DashboardSkeleton/index.tsx
@@ -1,9 +1,9 @@
 import { Center } from '@components/common';
 import { Desktop, Mobile, Tablet } from '@utils/responsive/Device';
 import type { DashboardProps } from '@utils/types/Dashboard';
-import { DesktopDashboardSkeleton } from './DesktopDashboardSkeleton';
-import { MobileDashboardSkeleton } from './MobileDashboardSkeleton';
-import { TabletDashboardSkeleton } from './TabletDashboardSkeleton';
+import { DesktopDashboardSkeleton } from './Desktop';
+import { MobileDashboardSkeleton } from './Mobile';
+import { TabletDashboardSkeleton } from './Tablet';
 
 type DashboardSkeletonProps = Omit<DashboardProps, 'contents'>;
