+++ app/src/components/templates/DashboardSkeleton/shared/DashboardSkeletonItem.tsx
